Moving data fetchers to core bundle
