+++ src/Sylius/Bundle/CoreBundle/DependencyInjection/SyliusCoreExtension.php
@@ -60,6 +60,13 @@ class SyliusCoreExtension extends AbstractResourceExtension implements PrependEx
         'templating',
         'twig',
         'mailer',
+        'reports'
+    );
+
+    private $emails = array(
+        'order_comment',
+        'order_confirmation',
+        'customer_welcome',
     );
 
     /**
