+++ src/Sylius/Bundle/WebBundle/Resources/views/Backend/Dashboard/main.html.twig
@@ -22,7 +22,7 @@
         <br>
     </div>
     <div class="col-md-6">
-        {{ render(controller('sylius.controller.report:embeddAction', {'report': 'ReportCode2'})) }}
+        {{ render(controller('sylius.controller.report:embeddAction', {'report': 'ReportCode'})) }}
     </div>
 </div>
